feat: moved ruint and garbleduint utilities
