+++ crates/interpreter/src/instructions/arithmetic.rs
@@ -1,38 +1,13 @@
 use core::ops::{Add, Div, Mul, Rem, Sub};
 
 use super::i256::{i256_div, i256_mod};
-use crate::{gas, Host, Interpreter};
-use compute::{self, uint::GarbledUint};
-use primitives::{ruint::Uint, U256};
+use crate::{
+    gas,
+    instructions::utility::{garbled_uint_to_ruint, ruint_to_garbled_uint},
+    Host, Interpreter,
+};
+use primitives::U256;
 use specification::hardfork::Spec;
-
-fn ruint_to_garbled_uint(value: &Uint<256, 4>) -> GarbledUint<256> {
-    let bytes: [u8; 32] = value.to_le_bytes();
-
-    let bits: Vec<bool> = bytes
-        .iter()
-        .flat_map(|&byte| (0..8).map(move |i| ((byte >> i) & 1) == 1))
-        .collect();
-
-    GarbledUint::<256>::new(bits)
-}
-
-fn garbled_uint_to_ruint(value: &GarbledUint<256>) -> Uint<256, 4> {
-    let bytes: Vec<u8> = value
-        .bits
-        .chunks(8)
-        .map(|chunk| {
-            chunk
-                .iter()
-                .enumerate()
-                .fold(0, |byte, (i, &bit)| byte | ((bit as u8) << i))
-        })
-        .collect();
-
-    let mut array = [0u8; 32];
-    array.copy_from_slice(&bytes);
-    Uint::from_le_bytes(array)
-}
 
 pub fn add<H: Host + ?Sized>(interpreter: &mut Interpreter, _host: &mut H) {
     gas!(interpreter, gas::VERYLOW);
@@ -168,21 +143,6 @@ mod tests {
         wiring::EthereumWiring<database_interface::EmptyDBTyped<core::convert::Infallible>, ()>,
     > {
         DummyHost::default()
-    }
-
-    #[test]
-    fn test_ruint_to_garbled_uint() {
-        let value = Uint::<256, 4>::from(123456789u64);
-        let garbled = ruint_to_garbled_uint(&value);
-        assert_eq!(garbled.bits.len(), 256);
-    }
-
-    #[test]
-    fn test_garbled_uint_to_ruint() {
-        let value = Uint::<256, 4>::from(123456789u64);
-        let garbled = ruint_to_garbled_uint(&value);
-        let result = garbled_uint_to_ruint(&garbled);
-        assert_eq!(value, result);
     }
 
     #[test]
